requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.2
sqlalchemy==2.0.23
asyncpg==0.29.0
gunicorn==21.2.0
openai==1.12.0
httpx==0.25.2
python-dotenv==1.0.0
aiohttp==3.9.1
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
psycopg2-binary==2.9.9
asyncpg
python-dotenv
pytest